+++ src/emmylua/repentanceapi.lua
@@ -2431,7 +2431,8 @@ ModCallbacks = {
 	MC_POST_ENTITY_KILL = 68, -- (Entity Ent)
 	MC_PRE_NPC_UPDATE = 69, -- (EntityNPC NPC), returns true if the internal ai should be ignored, false or nil/nothing otherwise
 	MC_PRE_SPAWN_CLEAN_AWARD = 70, -- (RNG& Rng, Vector SpawnPos), returns true if the spawn routine should be ignored, false or nil/nothing otherwise
-	MC_PRE_ROOM_ENTITY_SPAWN = 71 -- (EntityType Type, integer Variant, integer SubType, integer GridIndex, integer Seed) - This is called when entering a new room, before spawning entities which are part its layout. Grid entities will also trigger this callback and their type will the same as the type used by the gridspawn command. Because of this, effects are assigned the type 999 instead of 1000 in this callback. Optional return: an array table with new values { Type, Variant, Subtype }. Returning a table will override any replacements that might naturally occur i.e. enemy variants.
+	MC_PRE_ROOM_ENTITY_SPAWN = 71, -- (EntityType Type, integer Variant, integer SubType, integer GridIndex, integer Seed) - This is called when entering a new room, before spawning entities which are part its layout. Grid entities will also trigger this callback and their type will the same as the type used by the gridspawn command. Because of this, effects are assigned the type 999 instead of 1000 in this callback. Optional return: an array table with new values { Type, Variant, Subtype }. Returning a table will override any replacements that might naturally occur i.e. enemy variants.
+	MC_PRE_ENTITY_DEVOLVE = 72, -- (Entity Ent), returns true if the internal devolving behavior should be ignored - When returning true, this callback is responsible for spawning the devolved entity and removing the original one.
 }
 
 ---@enum EntityType
@@ -5002,7 +5003,25 @@ SoundEffect = {
 	SOUND_BEAST_GROWL = 815,
 	SOUND_BEAST_GRUMBLE = 816,
 	
-	NUM_SOUND_EFFECTS = 817
+	SOUND_FAMINE_GRUNT = 817,
+
+	SOUND_GFUEL_1 = 818,
+	SOUND_GFUEL_2 = 819,
+	SOUND_GFUEL_3 = 820,
+	SOUND_GFUEL_4 = 821,
+	SOUND_GFUEL_EXPLOSION = 822,
+	SOUND_GFUEL_EXPLOSION_BIG = 823,
+	SOUND_GFUEL_GUNSHOT = 824,
+	SOUND_GFUEL_GUNSHOT_SMALL = 825,
+	SOUND_GFUEL_GUNSHOT_LARGE = 826,
+	SOUND_GFUEL_GUNSHOT_SPREAD = 827,
+	SOUND_GFUEL_AIR_HORN = 828,
+	SOUND_GFUEL_ITEM_APPEAR = 829,
+	SOUND_GFUEL_GUNSHOT_MINI = 830,
+	SOUND_GFUEL_RICOCHET = 831,
+	SOUND_GFUEL_ROCKETLAUNCHER = 832,
+	
+	NUM_SOUND_EFFECTS = 833
 }
 
 ---@enum DoorState
@@ -6404,10 +6423,24 @@ TearFlags = {
 	TEAR_CARD_DROP_DEATH = TEARFLAG(79),			-- Killed enemies will drop a random tarot card
 	TEAR_RUNE_DROP_DEATH = TEARFLAG(80),			-- Killed enemies will drop a random rune
 	TEAR_TELEPORT = TEARFLAG(81),					-- Hit enemies will teleport to a different part of the room
+	TEAR_DECELERATE = TEARFLAG(82),					-- Decelerate over time
+	TEAR_ACCELERATE = TEARFLAG(83),					-- Accelerate over time
 	
-	TEAR_EFFECT_COUNT = 82,
+	TEAR_EFFECT_COUNT = 84,
 	
 	-- Reserved flags - cannot be randomly picked
+	TEAR_BOUNCE_WALLSONLY = TEARFLAG(104),			-- Similar to TEAR_BOUNCE but only bounces off walls, not enemies
+	TEAR_NO_GRID_DAMAGE = TEARFLAG(105),			-- Cannot deal damage to grid entities (used by Saturnus to prevent unfair damage in some rooms)
+	TEAR_BACKSTAB = TEARFLAG(106),					-- Deals extra damage from behind and inflicts bleeding
+	TEAR_FETUS_SWORD = TEARFLAG(107),				-- Fetuses whack their target with a sword and perform spin attacks
+	TEAR_FETUS_BONE = TEARFLAG(108),				-- Fetuses whack their target with a bone club instead of ramming into them
+	TEAR_FETUS_KNIFE = TEARFLAG(109),				-- Fetuses carry a knife
+	TEAR_FETUS_TECHX = TEARFLAG(110),				-- Fetuses have a Tech X ring around them
+	TEAR_FETUS_TECH = TEARFLAG(111),				-- Fetuses keep their distance and occasionally shoot tech lasers at their target
+	TEAR_FETUS_BRIMSTONE = TEARFLAG(112),			-- Fetuses shoot a brimstone beam at the first enemy they hit
+	TEAR_FETUS_BOMBER = TEARFLAG(113),				-- Fetuses drop a bomb on their first impact with an enemy
+	TEAR_FETUS = TEARFLAG(114),						-- Base flag for C Section fetuses
+
 	TEAR_REROLL_ROCK_WISP = TEARFLAG(115),			
 	TEAR_MOM_STOMP_WISP = TEARFLAG(116),			
 	TEAR_ENEMY_TO_WISP = TEARFLAG(117),				
@@ -6674,7 +6707,8 @@ SeedEffect = {
 	SEED_AXIS_ALIGNED_CONTROLS = 76,
 	SEED_SUPER_HOT = 77,
 	SEED_RETRO_VISION = 78,
-	NUM_SEEDS = 79
+	SEED_G_FUEL = 79,
+	NUM_SEEDS = 80
 }
 
 ---@enum GridRooms
@@ -8004,6 +8038,14 @@ end
 function HUD:SetVisible(Visible)
 end
 
+function HUD:Update()
+end
+
+function HUD:PostUpdate()
+end
+
+function HUD:Render()
+end
 
 
 ---@class Input